requirements.txt
# GANomics Project Dependencies
# Supporting GANomics

# Core Scientific Stack
numpy>=1.20.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep Learning Framework
torch>=1.10.0

# Data Normalization & Processing
tqdm>=4.62.0
pyyaml>=6.0

# Visualization (for scripts/plot.py and results/figures/)
matplotlib>=3.4.0
seaborn>=0.11.0

